added previous,next and play/pause buttons but play/pause is buggy, progress bar is also not working for some reason, might fix later taking a break

+++ music.js
@@ -4,6 +4,7 @@ import {readdir} from 'fs/promises';
 import ffmpeg from 'fluent-ffmpeg'
 import { EventEmitter } from 'events';
 import NodeID3 from 'node-id3';
+import { playBackEmitter } from './app.js';
 
 let currentSpeaker=null;
 let currentFFmpegCommand=null;
@@ -11,32 +12,61 @@ let currentFFmpegCommand=null;
 export const audioEvents = new EventEmitter();
 export const songEndEmitter=new EventEmitter()
 
-
+let nowPlayingPath, nowTime;
 export async function convertMP3toPCM(mp3Path,timeStamp=0,) {
-    let startTime = new Date();
 
     if (currentFFmpegCommand) {
         currentSpeaker.destroy();
-    currentFFmpegCommand.kill(currentFFmpegCommand.pid,'SIGTERM');
+        currentFFmpegCommand.kill(currentFFmpegCommand.pid,'SIGTERM');
         currentFFmpegCommand = null;
     }
 
+   nowPlayingPath=mp3Path;
+   nowTime=timeStamp;
+    
+    
+        playBackEmitter.once('play',()=>{
+            // console.log(nowTime)
+            playMusic(nowPlayingPath,timeStamp)
+        })
+        playBackEmitter.on('pause',()=>{
+            console.log('playBackEmitter pause')
+            if (currentFFmpegCommand) {
+            currentSpeaker.destroy();
+            currentFFmpegCommand.kill(currentFFmpegCommand.pid,'SIGTERM');
+            currentFFmpegCommand = null;
+            }
+        })
+
+    playMusic(mp3Path,timeStamp)
+    
+
+
+}
+
+function playMusic(mp3Path,timeStamp){
+    if (currentFFmpegCommand) {
+        currentSpeaker.destroy();
+        currentFFmpegCommand.kill(currentFFmpegCommand.pid,'SIGTERM');
+        currentFFmpegCommand = null;
+    }
+    console.log('something played this is playMusic()')
     try {
         const speaker = new Speaker();
         currentSpeaker=speaker;
 
-        const command = ffmpeg(mp3Path)
+        let command = ffmpeg(mp3Path)
         .setStartTime(timeStamp)
         .audioChannels(2)
         .audioFrequency(44100)
         .format('s16le')
          .on('start', () => {
-             // Capture start timestamp
+             playBackEmitter.emit('start')
          })
         .on('progress', (progress) => {
             
             // let timeElapsed = progress.timemark;
-            let timeElapsed=[Math.floor((startTime.getTime()%1000)/60),Math.floor(startTime.getTime()%1000)]
+
             
             const percentElapsed=progress.percent
             let timeMark=progress.timemark
@@ -59,17 +89,11 @@ export async function convertMP3toPCM(mp3Path,timeStamp=0,) {
 
         currentFFmpegCommand=command;
         command.pipe(currentSpeaker, { end: true });
-        
 
     } catch (err) {
         return true
     }
-    
-
-
 }
-
-
 
 import path from 'path';
 import os from 'os';
@@ -127,9 +151,9 @@ export async function getMusicFilesDataArray(musicFilesPath){
                 metadata.title=path.parse(filePath);
                 metadata.title=metadata.title.name;
         }
-        if(metadata.image===undefined){
-            metadata.image={imageBuffer:'./Pictures/2.png'}
-        }
+        // if(metadata.image===undefined){
+        //     metadata.image={imageBuffer:'./Pictures/2.png'}
+        // }
         if(metadata.length===undefined){
             metadata.length= await getAudioDuration(filePath)
             if(typeof metadata.length === 'number'){
@@ -144,7 +168,8 @@ export async function getMusicFilesDataArray(musicFilesPath){
             metadata.album,
             metadata.length,
             filePath,
-            metadata.image.imageBuffer];
+            // metadata.image.imageBuffer
+        ];
 
         row=row.map((element)=>{
             if(element===undefined){
